# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface Student {
  id: string;
  fullName: string;
  rollNo: number;
  className: string;
  section: string;
  stream?: string | null;
  additionalSubject?: string | null;
  gender?: string | null;
  dateOfBirth?: string | null;
  parentName?: string | null;
  parentMobile?: string | null;
  address?: string | null;
  email?: string | null;
  phone?: string | null;
  createdAt: string;
}

export interface CreateStudentInput {
  fullName: string;
  rollNo: number;
  className: string;
  section: string;
  stream?: string;
  additionalSubject?: string;
  gender?: string;
  dateOfBirth?: string;
  parentName?: string;
  parentMobile?: string;
  address?: string;
  email?: string;
  phone?: string;
}

export interface UpdateStudentInput {
  fullName?: string;
  rollNo?: number;
  className?: string;
  section?: string;
  stream?: string;
  additionalSubject?: string;
  gender?: string;
  dateOfBirth?: string;
  parentName?: string;
  parentMobile?: string;
  address?: string;
  email?: string;
  phone?: string;
}

export interface Teacher {
  id: string;
  fullName: string;
  subject: string;
  email?: string | null;
  phone?: string | null;
  assignedClasses?: string[] | null;
  createdAt: string;
}

export interface CreateTeacherInput {
  fullName: string;
  subject: string;
  email?: string;
  phone?: string;
  assignedClasses?: string[];
}

export interface UpdateTeacherInput {
  fullName?: string;
  subject?: string;
  email?: string;
  phone?: string;
  assignedClasses?: string[];
}

export interface SchoolClass {
  id: string;
  name: string;
  section: string;
  teacherId?: string | null;
  studentCount?: number | null;
  createdAt: string;
}

export interface CreateClassInput {
  name: string;
  section: string;
  teacherId?: string;
  studentCount?: number;
}

export interface UpdateClassInput {
  name?: string;
  section?: string;
  teacherId?: string;
  studentCount?: number;
}

export type ListStudentsParams = {
search?: string;
className?: string;
section?: string;
};

export type ListTeachersParams = {
search?: string;
};