# lib/api-zod/src/generated/types/createClassInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClassInput {
  name: string;
  section: string;
  teacherId?: string;
  studentCount?: number;
}