# lib/api-zod/src/generated/types/createStudentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentInput {
  fullName: string;
  rollNo: number;
  className: string;
  section: string;
  stream?: string;
  additionalSubject?: string;
  gender?: string;
  dateOfBirth?: string;
  parentName?: string;
  parentMobile?: string;
  address?: string;
  email?: string;
  phone?: string;
}