# lib/api-zod/src/generated/types/createTeacherInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeacherInput {
  fullName: string;
  subject: string;
  email?: string;
  phone?: string;
  assignedClasses?: string[];
}