# lib/api-zod/src/generated/types/schoolClass.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SchoolClass {
  id: string;
  name: string;
  section: string;
  teacherId?: string | null;
  studentCount?: number | null;
  createdAt: string;
}