# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: string;
  fullName: string;
  rollNo: number;
  className: string;
  section: string;
  stream?: string | null;
  additionalSubject?: string | null;
  gender?: string | null;
  dateOfBirth?: string | null;
  parentName?: string | null;
  parentMobile?: string | null;
  address?: string | null;
  email?: string | null;
  phone?: string | null;
  createdAt: string;
}