# lib/api-zod/src/generated/types/teacher.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Teacher {
  id: string;
  fullName: string;
  subject: string;
  email?: string | null;
  phone?: string | null;
  assignedClasses?: string[] | null;
  createdAt: string;
}