# lib/api-zod/src/generated/types/updateClassInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClassInput {
  name?: string;
  section?: string;
  teacherId?: string;
  studentCount?: number;
}