# lib/db/src/schema/teachers.ts
import { pgTable, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const teachersTable = pgTable("teachers", {
  id: text("id").primaryKey(),
  fullName: text("full_name").notNull(),
  subject: text("subject").notNull(),
  email: text("email"),
  phone: text("phone"),
  assignedClasses: text("assigned_classes").array().default([]),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertTeacherSchema = createInsertSchema(teachersTable);
export type InsertTeacher = z.infer<typeof insertTeacherSchema>;
export type Teacher = typeof teachersTable.$inferSelect;